+++ tools/vgsasm/src/offset.hpp
@@ -22,6 +22,11 @@ void offset_parse(LineData* line)
                 it++;
                 if (it == line->token.end() || it->first != TokenType::Other) {
                     line->error = true;
+                    line->errmsg = "No structure name specified in offset syntax.";
+                    return;
+                }
+                if (-1 == it->second.find('.')) {
+                    line->error = true;
                     line->errmsg = "No structure field name specified in offset syntax.";
                     return;
                 }
@@ -42,22 +47,25 @@ void offset_replace(LineData* line)
 {
     for (auto it = line->token.begin(); it != line->token.end(); it++) {
         if (it->first == TokenType::Offset) {
-            auto tokens = split_token(it->second, '.');
-            if (2 != tokens.size()) {
+            auto dot = it->second.find(".");
+            if (-1 == dot) {
+                // NOTE: this error case will not pass
                 line->error = true;
                 line->errmsg = "No structure field name specified in `offset`: " + it->second;
                 return;
             }
-            auto s = structTable.find(tokens[0]);
+            auto name = it->second.substr(0, dot);
+            auto field = it->second.substr(dot + 1);
+            auto s = structTable.find(name);
             if (s == structTable.end()) {
                 line->error = true;
-                line->errmsg = "Undefined structure " + it->second + " is specified in offset.";
+                line->errmsg = "Undefined structure " + name + " is specified in offset.";
                 return;
             }
             int offset = 0;
             bool found = false;
             for (auto f : s->second->fields) {
-                if (f->name == tokens[1]) {
+                if (f->name == field) {
                     it->first = TokenType::Numeric;
                     it->second = std::to_string(offset);
                     found = true;
@@ -68,7 +76,7 @@ void offset_replace(LineData* line)
             }
             if (!found) {
                 line->error = true;
-                line->errmsg = "Field name `" + tokens[1] + "` is not defined in structure `" + tokens[0] + "`.";
+                line->errmsg = "Field name `" + field + "` is not defined in structure `" + name + "`.";
                 return;
             }
         }