+++ tools/vgsasm/src/org.hpp
@@ -6,37 +6,52 @@
 #pragma once
 #include "common.h"
 
-static int lastOrg = -1;
+int lastOrg = -1;
 
 void org_parse(LineData* line)
 {
-    if (line->token.empty()) {
+    for (auto it = line->token.begin(); it != line->token.end(); it++) {
+        if (it->second != "ORG") {
+            continue;
+        }
+        if (it != line->token.begin()) {
+            line->error = true;
+            line->errmsg = "`org` must appear at the beginning of the line.";
+            return;
+        }
+        it++;
+        if (it == line->token.end()) {
+            line->error = true;
+            line->errmsg = "`org` address is not specified.";
+            return;
+        }
+        if (it->first != TokenType::Numeric) {
+            line->error = true;
+            line->errmsg = "Unexpected symbol: " + it->second;
+            return;
+        }
+        auto org = atoi(it->second.c_str());
+        if (0xFFFF < org || org < 0) {
+            line->error = true;
+            line->errmsg = "`org` address is incorrect: " + it->second;
+            return;
+        }
+        if (it + 1 != line->token.end()) {
+            line->error = true;
+            line->errmsg = "Unexpected symbol: " + (it + 1)->second;
+            return;
+        }
+        if (lastOrg >= org) {
+            line->error = true;
+            line->errmsg = "Multiple orgs must be written in ascending order from the beginning of the line.";
+            return;
+        }
+        lastOrg = org;
+        line->token[0].first = TokenType::Org;
+        line->token[0].second = line->token[1].second;
+        line->token.erase(it);
         return;
     }
-    if (line->token[0].second != "ORG") {
-        return;
-    }
-    if (line->token.size() != 2 || line->token[1].first != TokenType::Numeric) {
-        line->error = true;
-        line->errmsg = "ORG formatting is incorrect.";
-        return;
-    }
-    auto org = atoi(line->token[1].second.c_str());
-    if (0xFFFF < org || org < 0) {
-        line->error = true;
-        line->errmsg = "org is not the correct address (0x0000 to 0xFFFF)";
-        return;
-    }
-    if (lastOrg >= org) {
-        line->error = true;
-        line->errmsg = "Multiple orgs must be written in ascending order from the beginning of the line.";
-        return;
-    }
-    lastOrg = org;
-
-    line->token[0].first = TokenType::Org;
-    line->token[0].second = line->token[1].second;
-    line->token.erase(line->token.begin() + 1);
 }
 
 void org_setpc(std::vector<LineData*>* lines)